v0.96
Added single player, loading screen & selection screen.
Added sound effects.

Wins on the UI aren't updated.

Bugs: Buttons are initially in the wrong place in single player.
Buttons are in the wrong position on new game.
Both of these bugs just need the position to be set as soon as the position is drawn.